Fix problema de ler arquivo que não existe
O arquivo application.rb poderia tentar ler o arquivo local_env.yml
sem este existir o que causava erro de failed to load file e
quebrava os tests no Travis.

+++ config/application.rb
@@ -14,12 +14,13 @@ module Zion
 
     config.before_configuration do
       env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      variables = YAML.load(File.open(env_file))
-
-      if variables
-        variables.each do |key, value|
-          ENV[key.to_s] = value
-        end if File.exists?(env_file)
+      if File.exists?(env_file)
+        variables = YAML.load(File.open(env_file))
+        if variables
+          variables.each do |key, value|
+            ENV[key.to_s] = value
+          end
+        end
       end
     end
